requirements.txt
numpy
opencv-python
Pillow
PyYAML
scikit-image
scikit-learn
torch
torchvision
tqdm
